Unso some premature changes
Local and Resident invoker both use _same_ parser, the
DefaultMavenParser that is reusable across all "just maven"
invokers.
