+++ maven-cli/src/main/java/org/apache/maven/cling/invoker/mvn/DefaultMavenParser.java
@@ -18,29 +18,25 @@
  */
 package org.apache.maven.cling.invoker.mvn;
 
-import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Stream;
 
+import org.apache.commons.cli.ParseException;
 import org.apache.maven.api.cli.Options;
 import org.apache.maven.api.cli.ParserException;
 import org.apache.maven.api.cli.ParserRequest;
 import org.apache.maven.api.cli.extensions.CoreExtension;
 import org.apache.maven.api.cli.mvn.MavenInvokerRequest;
 import org.apache.maven.api.cli.mvn.MavenOptions;
-import org.apache.maven.api.cli.mvn.MavenParser;
-import org.apache.maven.cling.invoker.BaseParser;
+import org.apache.maven.api.cli.mvn.local.LocalMavenParser;
 
-public abstract class DefaultMavenParser<O extends MavenOptions, R extends MavenInvokerRequest<O>>
-        extends BaseParser<O, R> implements MavenParser<R> {
+public class DefaultMavenParser extends BaseMavenParser<MavenOptions, MavenInvokerRequest<MavenOptions>>
+        implements LocalMavenParser {
     @SuppressWarnings("ParameterNumber")
     @Override
-    protected abstract R getInvokerRequest(
+    protected DefaultMavenInvokerRequest<MavenOptions> getInvokerRequest(
             ParserRequest parserRequest,
             Path cwd,
             Path installationDirectory,
@@ -50,38 +46,34 @@ public abstract class DefaultMavenParser<O extends MavenOptions, R extends Maven
             Path topDirectory,
             Path rootDirectory,
             ArrayList<CoreExtension> extensions,
-            Options options);
+            Options options) {
+        return new DefaultMavenInvokerRequest<>(
+                parserRequest,
+                cwd,
+                installationDirectory,
+                userHomeDirectory,
+                userProperties,
+                systemProperties,
+                topDirectory,
+                rootDirectory,
+                parserRequest.in(),
+                parserRequest.out(),
+                parserRequest.err(),
+                extensions,
+                (MavenOptions) options);
+    }
 
     @Override
-    protected List<O> parseCliOptions(Path rootDirectory, List<String> args) throws ParserException, IOException {
-        ArrayList<O> result = new ArrayList<>();
-        // CLI args
-        result.add(parseMavenCliOptions(args));
-        // maven.config; if exists
-        Path mavenConfig = rootDirectory.resolve(".mvn/maven.config");
-        if (Files.isRegularFile(mavenConfig)) {
-            result.add(parseMavenConfigOptions(mavenConfig));
-        }
-        return result;
-    }
-
-    protected O parseMavenCliOptions(List<String> args) throws ParserException {
-        return parseArgs(Options.SOURCE_CLI, args);
-    }
-
-    protected O parseMavenConfigOptions(Path configFile) throws ParserException, IOException {
-        try (Stream<String> lines = Files.lines(configFile, Charset.defaultCharset())) {
-            List<String> args =
-                    lines.filter(arg -> !arg.isEmpty() && !arg.startsWith("#")).toList();
-            O options = parseArgs("maven.config", args);
-            if (options.goals().isPresent()) {
-                // This file can only contain options, not args (goals or phases)
-                throw new ParserException("Unrecognized maven.config file entries: "
-                        + options.goals().get());
-            }
-            return options;
+    protected MavenOptions parseArgs(String source, List<String> args) throws ParserException {
+        try {
+            return CommonsCliMavenOptions.parse(source, args.toArray(new String[0]));
+        } catch (ParseException e) {
+            throw new ParserException("Failed to parse source " + source, e.getCause());
         }
     }
 
-    protected abstract O parseArgs(String source, List<String> args) throws ParserException;
+    @Override
+    protected MavenOptions assembleOptions(List<MavenOptions> parsedOptions) {
+        return LayeredMavenOptions.layerMavenOptions(parsedOptions);
+    }
 }
