+++ maven-cli/src/test/java/org/apache/maven/cling/invoker/mvn/resident/DefaultResidentMavenInvokerTest.java
@@ -26,9 +26,10 @@ import com.google.common.jimfs.Configuration;
 import com.google.common.jimfs.Jimfs;
 import org.apache.maven.api.cli.Invoker;
 import org.apache.maven.api.cli.Parser;
-import org.apache.maven.api.cli.mvn.resident.ResidentMavenInvokerRequest;
-import org.apache.maven.api.cli.mvn.resident.ResidentMavenOptions;
+import org.apache.maven.api.cli.mvn.MavenInvokerRequest;
+import org.apache.maven.api.cli.mvn.MavenOptions;
 import org.apache.maven.cling.invoker.ProtoLookup;
+import org.apache.maven.cling.invoker.mvn.DefaultMavenParser;
 import org.apache.maven.cling.invoker.mvn.MavenInvokerTestSupport;
 import org.codehaus.plexus.classworlds.ClassWorld;
 import org.junit.jupiter.api.Disabled;
@@ -40,18 +41,18 @@ import org.junit.jupiter.api.io.TempDir;
  * Resident UT.
  */
 public class DefaultResidentMavenInvokerTest
-        extends MavenInvokerTestSupport<ResidentMavenOptions, ResidentMavenInvokerRequest> {
+        extends MavenInvokerTestSupport<MavenOptions, MavenInvokerRequest<MavenOptions>> {
 
     @Override
-    protected Invoker<ResidentMavenInvokerRequest> createInvoker() {
+    protected Invoker<MavenInvokerRequest<MavenOptions>> createInvoker() {
         return new DefaultResidentMavenInvoker(ProtoLookup.builder()
                 .addMapping(ClassWorld.class, new ClassWorld("plexus.core", ClassLoader.getSystemClassLoader()))
                 .build());
     }
 
     @Override
-    protected Parser<ResidentMavenInvokerRequest> createParser() {
-        return new DefaultResidentMavenParser();
+    protected Parser<MavenInvokerRequest<MavenOptions>> createParser() {
+        return new DefaultMavenParser();
     }
 
     @Test
